refactor: move exception class to application
